refactor input and data storage structure
This changes data organisation to assume that by default all data used
by lagtraj is stored in `data/` relative to where lagtraj is being run
from. Input files which are distributed with lagtraj are available by
passing in `lagtraj://<input_name>` for accessing for example domain
definitions included in lagtraj.