Check local input defns when using bundled ones
When a user tries to use an input-definition by default we copy that
input definition to the user's local `data/` directory. We now check
if there is an existing file there and compare these, if there are any
differences we halt the operation and ask the user to either delete the
file or request to use the local file explicitly instead.