Ensure domain data creation time is fixed (#105)
Previously domain data creation time was set to current time every time
domain data is loaded. This doesn't work when attempting to apply
conversion to an existing forcing file since the attributes of the
loaded forcing file don't match the attributes derived a conversion time
(since conversion might happen as a separate and later step). The
"creation time" for domain data is now set to the creation time for the
first downloaded file. This ensures that the creation time doesn't
change as domain data is downloaded and only changes if all domain data
is deleted and re-downloaded.