Fix inconsistency for Eulerian traj without domain
Related to https://github.com/EUREC4A-UK/lagtraj/issues/112

For Eulerian trajectories we either define a fixed timestep (e.g.
`PT1H`) and don't need domain data (and so the domain shouldn't be
given) or we are using timesteps from the domain data (in which case the
domain to use needs to be defined)

This introduces a special flag `__is_set__` for variables in input
definitions in that require that other variables are set.