Fix readme typos with cleaned examples (#143)
Clean up example input files to fit with what we're using in the EUREC4A campaign. Ensure that all examples mentioned in the README work by adding tests that parse the README and tries to run each command. Also implements functionality to check if all data is available to run a given command and updates the test data to match the new included examples.

Co-authored-by: Steven Boeing <sboeing@host596.jc.rl.ac.uk>